Add sending content as JSON to conversation
