+++ demos/src/Examples/Messaging/content.js
@@ -1,3 +1,3 @@
 export const content = `
-  <p><em>This is <s>a extensive</s> a simple </em><strong><em>messaging</em></strong><em> </em><strong><em>example</em></strong><em>.</em></p><p>You can add <em>images</em>, <em>bullet lists</em>, <em>ordered lists</em>, <em>blockquotes</em> or <code>code</code> here.</p><p>Add emojis like this <span data-name="zap" data-type="emoji" contenteditable="false">⚡</span> by typing <code>:</code>-symbol or click the <code>set emoji</code> or <code>select emoji</code> button.</p><p>Also you can add mentions like this <span data-type="mention" class="mention" data-id="Lea Thompson" contenteditable="false">@Lea Thompson</span> by typing <code>@</code>-symbol or click the <code>set mention</code> or <code>select mention</code> button.</p><p>In addition you can send the message by clicking <strong><em>send button</em></strong> or hitting <code>cmd+enter</code> on mac or <code>ctrl+enter</code> on pc.</p>
+  <p><em>This is <s>a extensive</s> a simple </em><strong><em>messaging</em></strong><em> </em><strong><em>example</em></strong><em> </em>with <code>@</code>-mentions <span data-type="mention" class="mention" data-id="Madonna" contenteditable="false">@Madonna</span> and <code>:</code>-emojis <span data-name="grin" data-type="emoji" contenteditable="false">😁</span> <em>.</em></p><p>In addition you can send the message by clicking <strong><em>send button</em></strong> or hitting <code>cmd+enter</code> on mac or <code>ctrl+enter</code> on pc.</p>
 `
